+++ goat_route/core/database/db_helpers.py
