--- goat_route/models/interfaces.py
@@ -1,4 +1,0 @@
-from utilities.observer import IObservable
-
-class IModel(IObservable):
-    pass
