+++ goat_route/mvc/models.py
@@ -1,25 +1,18 @@
 from os import getenv
+from ctkmvc.model import ObservableModel
 
-from .interfaces import IModel
-from utilities.observer import ModelObserver
-from utilities.api_clients.yandex_api_clients import HTTPGeocoderClient
-from utilities.tsp_solver import TSPSolver
-from utilities.mapping.geolocation import Address, Coordinate
-from utilities.mapping.map import Map
+from core.api_clients.yandex_api_clients import HTTPGeocoderClient
+from core.tsp_solver import TSPSolver
+from core.mapping.geolocation import Address, Coordinate
+from core.mapping.map import Map
 
-class MainWindowModel(IModel):
+
+class MainWindowModel(ObservableModel):
 
     def __init__(self) -> None:
 
         super().__init__()
 
-        self._observers = []
-
-    def add_observer(self, observer: ModelObserver):
-        self._observers.append(observer)
-
-    def remove_observer(self, observer: ModelObserver):
-        self._observers.remove(observer)
 
     def _get_coordinates(self, addresses):
 
@@ -38,11 +31,13 @@ class MainWindowModel(IModel):
 
         return coords
     
+
     def _solve_tsp(self, address: Address, coords: list[Coordinate], network_type):
 
         tsp_solver = TSPSolver()
 
         return tsp_solver.solve(address, coords, network_type)
+    
     
     def _save_html_map(self, solution):
 
@@ -63,10 +58,3 @@ class MainWindowModel(IModel):
         solution = self._solve_tsp(Address.convert_from(addresses[0]), coords, network_type)
 
         self._save_html_map(solution)
-
-
-    def notify_observers(self):
-
-        for observer in self._observers:
-            
-            observer.model_is_changed()
