+++ goat_route/resources/resource.py
@@ -1,0 +1,7 @@
+from enum import StrEnum
+
+class Resource(StrEnum):
+
+    ICON = 'resources\\icon.ico'
+    LOGO = 'resources\\logo.png'
+    MAP = 'resources\\map.html'
